{{cookiecutter.project_slug}}/requirements.txt
Flask==0.11.0
flask-sqlalchemy==2.1
py-kim==1.2.0
sqlalchemy-utils==0.32.13
arrow==0.10.0
Flask-Alembic==2.0.1
ipdb
psycopg2==2.7.1
pytest==3.1.2
pytest-cov==2.5.1
PyJWT==1.5.2
gunicorn==19.3.0
factory_boy==2.8.1
freezegun==0.3.9
requests==2.18.1
behave==1.2.5
behave_http==0.1.1
arrested==0.1.1
passlib
